Major rework on MaterialWidget test cases - Refactor base test classes.
