+++ gwt-material/src/test/java/gwt/material/design/client/ui/MaterialFABTest.java
@@ -21,7 +21,7 @@ package gwt.material.design.client.ui;
 
 import com.google.gwt.user.client.ui.Widget;
 import gwt.material.design.client.constants.*;
-import gwt.material.design.client.ui.base.MaterialWidgetTest;
+import gwt.material.design.client.ui.base.MaterialWidgetTestCase;
 import gwt.material.design.client.ui.html.ListItem;
 
 /**
@@ -30,7 +30,7 @@ import gwt.material.design.client.ui.html.ListItem;
  * @author kevzlou7979
  * @author Ben Dol
  */
-public class MaterialFABTest extends MaterialWidgetTest<MaterialFAB> {
+public class MaterialFABTest extends MaterialWidgetTestCase<MaterialFAB> {
 
     static final int FAB_ITEM_SIZE = 5;
