+++ gwt-material/src/test/java/gwt/material/design/client/ui/base/BaseTestCase.java
@@ -27,10 +27,12 @@ import gwt.material.design.client.resources.MaterialResources;
 import gwt.material.design.client.resources.WithJQueryResources;
 import org.junit.Ignore;
 
+import java.util.logging.Logger;
+
 import static gwt.material.design.jquery.client.api.JQuery.$;
 
 @Ignore
-public class MaterialTestCase extends GWTTestCase {
+public class BaseTestCase extends GWTTestCase {
 
     private boolean tearDownRootPanel = false;
